requirements.txt
tensorflow==2.8
pillow
streamlit
mediapipe
opencv-python